refactor(mutants): move inline test mods to sibling *_tests.rs files
The new inline tests added in this PR push SonarCloud's new-code
duplication density to 7.5% (gate is 3%). The repo already excludes
`**/*_tests.rs` from CPD per `sonar-project.properties:14` and uses
sibling test files (e.g. `shared_tests.rs`, `list_tests.rs`,
`bug_tests.rs`) wired via `#[cfg(test)] #[path = "X_tests.rs"] mod tests;`.

Move the inline `#[cfg(test)] mod tests` blocks in the three flagged
files to siblings:

- `src/client/bug.rs`     -> `src/client/bug_tests.rs`     (715 lines)
- `src/client/group.rs`   -> `src/client/group_tests.rs`   (530 lines)
- `src/commands/field.rs` -> `src/commands/field_tests.rs` (114 lines)

Module structure is preserved (still `client::bug::tests` etc.) so all
`super::*` / `super::super::*` paths inside the test bodies resolve
unchanged. No test logic, no helpers, no abstractions changed.

Per `docs/specs/2026-04-27-sonar-refactor-design.md:175`, extracting
shared test helpers to dedupe these blocks would create bad abstractions
over genuinely-distinct test scenarios; the file move is the convention
the repo already follows.
